rag_backend/requirements.txt
fastapi
uvicorn
openai
pydantic
python-dotenv
scikit-learn
jieba
numpy
chromadb
neo4j
